src/requirements.txt
# Used by SAM build when packaging Lambda (copy from root or keep in sync)
awslabs.mcp-lambda-handler>=0.1.12
